Moving DependencyGraph out of Substitutions
