+++ tests/CMakeLists.txt
@@ -25,6 +25,7 @@ addUnitTest(AffineExprTest)
 addUnitTest(AssignmentTest)
 addUnitTest(CompiledAssignmentTest)
 addUnitTest(ConstraintTest)
+addUnitTest(DependencyGraph)
 addUnitTest(LeastSquareSolverTest SolverTestFunctions.cpp)
 addUnitTest(LinearizedTaskConstraintTest SolverTestFunctions.cpp)
 addUnitTest(MatrixPropertiesTest)
