+++ tests/SubstitutionTest.cpp
@@ -24,123 +24,6 @@ using namespace tvm::hint;
 using namespace tvm::hint::internal;
 using namespace Eigen;
 
-struct compVec
-{
-  bool operator()(const std::vector<size_t>& u, const std::vector<size_t>& v) const
-  {
-    if (u.size() < v.size())
-      return true;
-    else if (v.size() < u.size())
-      return false;
-    else
-      return std::set<size_t>(u.begin(), u.end()) < std::set<size_t>(v.begin(), v.end());
-  }
-};
-
-bool equal(const std::vector<std::vector<size_t>>& u, const std::vector<std::vector<size_t>>& v)
-{
-  if (u.size() == v.size())
-  {
-    std::set<std::vector<size_t>, compVec> su(u.begin(), u.end());
-    std::set<std::vector<size_t>, compVec> sv(v.begin(), v.end());
-    //using ==, <=, ... between two sets seems to always use ==, <=, ... on the
-    //elements of the set, irrespective of the comparator given to the set. We
-    //thus relies on std::lexicographical_compare with our custom comparator
-    return !std::lexicographical_compare(su.begin(), su.end(), sv.begin(), sv.end(), compVec())
-        && !std::lexicographical_compare(sv.begin(), sv.end(), su.begin(), su.end(), compVec());
-  }
-  else
-    return false;
-}
-
-namespace tvm
-{
-namespace hint
-{
-namespace internal
-{
-
-  class SubstitutionTest
-  {
-  public:
-    static void testSCC()
-    {
-      internal::Substitutions::DependencyGraph g;
-      g.addNode();
-      g.addNode();
-      g.addNode();
-      g.addNode();
-      g.addNode();
-      g.addNode();
-      g.addEdge(0, 3);
-      g.addEdge(3, 1);
-      g.addEdge(1, 0);
-      g.addEdge(1, 2);
-      g.addEdge(1, 4);
-      g.addEdge(2, 5);
-      g.addEdge(5, 2);
-
-      auto r = g.reduce();
-      const auto& e = r.second.edges();
-
-      FAST_CHECK_UNARY(equal(r.first, { { 0,1,3 },{ 4 },{ 2,5 } }));
-      FAST_CHECK_EQ(e.size(), 2);
-      //we have two edges, one from the SCC with size 3 to the SCC with size 1,
-      //one from the SCC with size 3 to the SCC with size 2
-      auto it1 = e.begin();
-      auto it2 = e.begin(); ++it2;
-      FAST_CHECK_EQ(r.first[it1->first].size(), 3);
-      FAST_CHECK_EQ(r.first[it2->first].size(), 3);
-      auto s1 = r.first[it1->second].size();
-      auto s2 = r.first[it2->second].size();
-      FAST_CHECK_UNARY((s1 == 1 && s2 == 2) || (s1 == 2 && s2 == 1));
-    }
-
-    static void testOrder()
-    {
-      internal::Substitutions::DependencyGraph g;
-      g.addNode();
-      g.addNode();
-      g.addNode();
-      g.addNode();
-      g.addNode();
-      g.addNode();
-      g.addNode();
-      g.addNode();
-      g.addNode();
-      g.addNode();
-      g.addEdge(0, 1);
-      g.addEdge(0, 2);
-      g.addEdge(1, 2);
-      g.addEdge(1, 4);
-      g.addEdge(2, 4);
-      g.addEdge(3, 5);
-      g.addEdge(3, 6);
-      g.addEdge(5, 6);
-      g.addEdge(8, 9);
-      g.addEdge(9, 4);
-
-      auto r = g.order();
-
-      std::set<std::vector<size_t>> v = { { 4,2,1,0,9,8 },{ 6,5,3 },{ 7 } };
-      FAST_CHECK_EQ(std::set<std::vector<size_t>>(r.begin(), r.end()), v);
-    }
-  };
-
-
-}
-}
-}
-
-TEST_CASE("SCC test")
-{
-  SubstitutionTest::testSCC();
-}
-
-TEST_CASE("Order test")
-{
-  SubstitutionTest::testOrder();
-}
 
 TEST_CASE("GenericCalculator")
 {
